Separated login and userregistration layout components and renamed it. Added usePost methods to layout components' handlers, but it needs fix.
