+++ frontend/layouts/SignUpLayout.tsx
@@ -1,10 +1,8 @@
 import { Button, Container, Grid, TextField } from "@mui/material";
 // eslint-disable-next-line import/named
 import Box from "@mui/material/Box";
-import Checkbox from "@mui/material/Checkbox";
 import CssBaseline from "@mui/material/CssBaseline";
 import FormControl from "@mui/material/FormControl";
-import FormControlLabel from "@mui/material/FormControlLabel";
 import InputLabel from "@mui/material/InputLabel";
 import MenuItem from "@mui/material/MenuItem";
 import Select, { SelectChangeEvent } from "@mui/material/Select";
@@ -16,83 +14,7 @@ import * as React from "react";
 
 //FIXME: page height!
 
-export const LoginLayout: React.FC<{ children: React.ReactNode }> = ({ children }) => {
-    const theme = useTheme();
-    const handleSubmit = (event: React.FormEvent<HTMLFormElement>) => {
-        event.preventDefault();
-        const data = new FormData(event.currentTarget);
-        console.log({
-            email: data.get("email"),
-            password: data.get("password"),
-        });
-    };
-
-    return (
-        <Container component="main" maxWidth="xs">
-            <CssBaseline />
-            <Box
-                component="form"
-                onSubmit={handleSubmit}
-                sx={{
-                    marginTop: 8,
-                    display: "flex",
-                    flexDirection: "column",
-                    alignItems: "center",
-                }}
-                noValidate
-                autoComplete="off"
-            >
-                <Typography component="h1" variant="h5">
-                    サインイン
-                </Typography>
-                <Box component="form" noValidate sx={{ mt: 8 }}></Box>
-                <TextField
-                    margin="normal"
-                    required
-                    fullWidth
-                    id="email"
-                    label="メール"
-                    name="email"
-                    autoComplete="email"
-                    autoFocus
-                />
-                <TextField
-                    margin="normal"
-                    required
-                    fullWidth
-                    name="password"
-                    label="パスワード"
-                    type="password"
-                    id="password"
-                    autoComplete="current-password"
-                />
-                <FormControlLabel
-                    control={<Checkbox value="remember" color="primary" />}
-                    label="ログイン情報保存"
-                />
-                <Button type="submit" fullWidth variant="contained" sx={{ mt: 3, mb: 2 }}>
-                    サインイン
-                </Button>
-                <Grid container>
-                    <Grid item xs>
-                        <Link href="#" variant="body2">
-                            パスワード失望
-                        </Link>
-                    </Grid>
-                    <Grid item>
-                        <Link href="/userregistration" variant="body2">
-                            {"新規登録"}
-                        </Link>
-                    </Grid>
-                </Grid>
-            </Box>
-        </Container>
-    );
-};
-
-export const UserRegisterLayout: React.FC<{ children: React.ReactNode }> = ({
-    children,
-}) => {
+export const SignUpLayout: React.FC<{ children: React.ReactNode }> = ({ children }) => {
     const handleSubmit = (event: React.FormEvent<HTMLFormElement>) => {
         event.preventDefault();
         const data = new FormData(event.currentTarget);
@@ -106,29 +28,29 @@ export const UserRegisterLayout: React.FC<{ children: React.ReactNode }> = ({
             password: data.get("password"),
         });
 
-        //     const { doPost, isLoading } = usePost({
-        //         method: "post",
-        //         url: "/api/user/",
-        //     });
-        //     doPost({
-        //         url: "/api/user/",
-        //         params: {
-        //             familyName: data.get("family_name"),
-        //             givenName: data.get("given_name"),
-        //             familyKana: data.get("family_kana"),
-        //             givenKana: data.get("given_kana"),
-        //             positionSelect: data.get("position-select"),
-        //             email: data.get("email"),
-        //             password: data.get("password"),
-        //         },
-        //         onSuccess: (familyName) => {
-        //             console.log(`create ${familyName}  success!`);
-        //         },
-        //         onError: (err) => {
-        //             console.log(err.message);
-        //         },
-        //     });
-        // };
+        //FIXME: usePost params, body
+        // const { doPost } = usePost({
+        //     method: "post",
+        //     url: "/api/user",
+        // });
+        // doPost({
+        //     url: "/api/user",
+        //     params: {
+        //         familyName: data.get("family_name"),
+        //         givenName: data.get("given_name"),
+        //         familyKana: data.get("family_kana"),
+        //         givenKana: data.get("given_kana"),
+        //         positionSelect: data.get("position-select"),
+        //         email: data.get("email"),
+        //         password: data.get("password"),
+        //     },
+        //     onSuccess: (familyName) => {
+        //         console.log(`create ${familyName}  success!`);
+        //     },
+        //     onError: (err) => {
+        //         console.log(err.message);
+        //     },
+        // });
     };
 
     const theme = useTheme();
@@ -232,7 +154,7 @@ export const UserRegisterLayout: React.FC<{ children: React.ReactNode }> = ({
                 </Button>
                 <Grid container>
                     <Grid item>
-                        <Link href="/login" variant="body2">
+                        <Link href="/signin" variant="body2">
                             {"←ログインへ戻る"}
                         </Link>
                     </Grid>
